Moved asset settings to its own class
