+++ Assets/Scripts/Settings/UISettingsAssets.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01ee252c6096dbc45a4d5bea9fc25274
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
